refactor: assets refactor
