+++ src/assets/main-page/mainPageImg.png
