+++ src/spa/view/footer/footer.scss
@@ -8,7 +8,7 @@
     &-rs-img {
         width: 10rem;
         height: 4rem;
-        @include background ($src: '@src/assets/rs_school.svg');
+        @include background ($src: '@src/assets/footer/rs_school.svg');
     }
     &-gh-img {
         width: 4rem;
@@ -29,13 +29,13 @@
 .footer-link {
     &:nth-child(odd) {
         .footer-gh-img {
-            @include background ($src: '@src/assets/github_1.svg');
+            @include background ($src: '@src/assets/footer/github_1.svg');
         }
     }
 
     &:nth-child(even) {
         .footer-gh-img {
-            @include background ($src: '@src/assets/github_2.svg');
+            @include background ($src: '@src/assets/footer/github_2.svg');
         }
     }
 }
