+++ src/spa/view/input/passwordInput/passwordInput.scss
@@ -20,11 +20,11 @@
   background-position: center;
 
   &_opened {
-    background-image: url("@src/assets/eye-open.png");
+    background-image: url("@src/assets/password/eye-open.png");
   }
 
   &_closed {
-    background-image: url("@src/assets/eye-close.png");
+    background-image: url("@src/assets/password/eye-close.png");
   }
 
   &:hover {
@@ -34,5 +34,4 @@
 
 .input_password {
   width: 100%;
-  // padding-right: 2rem; // as here // chenged by Kushal
 }
