+++ src/spa/view/pages/productPage/productPage.scss
@@ -34,7 +34,7 @@
   left: 0.5rem;
   width: 10rem;
   aspect-ratio: 1/1.3;
-  @include background ($src: '@src/assets/sale.png');
+  @include background ($src: '@src/assets/product-page/sale.png');
 }
 
 .product__prices-wrapper {
